+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/wagner/opmodes/mechanismtests/TestArm.java
@@ -1,41 +1,50 @@
 package org.firstinspires.ftc.teamcode.wagner.opmodes.mechanismtests;
 
 import com.acmerobotics.roadrunner.PoseVelocity2d;
-import com.acmerobotics.roadrunner.Vector2d;
-import com.acmerobotics.roadrunner.ftc.Actions;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.MecanumDrive;
 import org.firstinspires.ftc.teamcode.wagner.GlobalStorage;
+import org.firstinspires.ftc.teamcode.wagner.mechanisms.Arm;
 import org.firstinspires.ftc.teamcode.wagner.mechanisms.Claw;
 import org.firstinspires.ftc.teamcode.wagner.nggamepad.NGGamepad;
 
-public class TestClaw extends LinearOpMode {
+public class TestArm extends LinearOpMode {
     @Override
     public void runOpMode() throws InterruptedException {
         MecanumDrive drive = new MecanumDrive(hardwareMap, GlobalStorage.pose);
 
         NGGamepad ng_gamepad1 = new NGGamepad(gamepad1);
-        Claw claw = new Claw();
-        claw.init(hardwareMap);
+        Arm arm = new Arm();
+        arm.init(hardwareMap);
 
         waitForStart();
 
-        telemetry.addData("Please wait", "Testing claw");
+        telemetry.addData("Please wait", "Testing arm");
+        telemetry.addData("Arm: Elbow", "Testing elbow");
         telemetry.update();
 
-        claw.openClaw(0);
-        test(1, claw.claw.getPosition() == 0, telemetry);
+        arm.setElbow(0.5);
+        test(1, arm.elbow.getPosition() == 0.5, telemetry);
 
-        claw.openClaw(1);
-        test(2, claw.claw.getPosition() == 1, telemetry);
+        arm.setElbow(0);
+        test(2, arm.elbow.getPosition() == 0, telemetry);
 
-        claw.openClaw(0.5);
-        test(3, claw.claw.getPosition() == 0.5, telemetry);
+        arm.setElbow(1);
+        test(2, arm.elbow.getPosition() == 1, telemetry);
 
-        claw.openClaw(Claw.clawOpenPosition);
-        test(4, claw.claw.getPosition() == Claw.clawOpenPosition, telemetry);
+        telemetry.addData("Arm: Shoulder", "Testing shoulder");
+        telemetry.update();
+
+        arm.setShoulder(0.5);
+        test(1, arm.shoulder.getPosition() == 0.5, telemetry);
+
+        arm.setShoulder(0);
+        test(2, arm.shoulder.getPosition() == 0, telemetry);
+
+        arm.setShoulder(1);
+        test(2, arm.shoulder.getPosition() == 1, telemetry);
 
         telemetry.addData("Tests", "Finished");
         telemetry.update();